refactoring countLoanedBook, getBookStatistics with using JPA queryMethod

+++ src/main/kotlin/com/group/libraryapp/dto/book/response/BookStatResponse.kt
@@ -4,7 +4,7 @@ import com.group.libraryapp.type.BookType
 
 data class BookStatResponse(
     val type: BookType,
-    var count: Int,
+    var count: Long,
 ) {
     fun plusOne() {
         count++
